Reloading update.
Inside ammo component.
Start reload with button press. If already reloading then stop the reload. If not start reload. Gets reload time of the weapon or a 3 second reload time if can't find a weapon since currently using player specific stuff so will eventually need to update for enemies.

Reload threshold for what percentage would increase current ammo. Reload percentage set based on current ammo divided by max ammo and then if above 0 reload time is updated to be the remainder time for how much ammo needs to be reloaded.

Reloading calculations are in the event tick of the ammo component with a tick interval of 0.01 so not frame dependant. Various calculations in here just check in ammo.

Cycling weapons will check if reloading and stop it if so.
Adding a new weapon to the player's weaponry will also check if the current weapon is valid and is reloading and if so will stop the reload.

Also implemented widget animations for getting ammo and using ammo just need audio to make it feel better. These widget animations are in the actual weapon widgets not the HUD.