Shoot location as a socket on each individual static mesh.
The shooting location of each gun is an arrow in the actor that has its location set based off of a socket in the weapon static mesh. This method will probably also be used for melee weapon hit boxes by using a socket in the static mesh where the hitbox should start and end. Could also use this method for the place where the static mesh should be for holding it but not changing that yet and instead just changing the pivot of the static meshes is enough.

Weapon and projectile names are now as enums.

Started working on shooting functionality as didnt want to do aiming animations.

Copied the manny animation blueprint and renamed to bipedal which will be used for both player and enemies that are bipedals. Will just need to make sure skeletons are similar enough.

Various checks in the animation blueprint event graph not going to detail them here but it enables aiming and stuff.

New state machine for aiming.
Cache this aiming pose. Then layered blend with the main states cached pose which leads into a blend by bool of aiming. If aiming will use the layered blend if not aiming then will use the main states cached pose.

Got loads of animations from mixamo using mixamo to unreal converter and created a revolver aim offset 1d and some animations for it just need to adjust them as see fit to make them look better when in use in first person. Might also need to do some stuff with the camera not sure yet.

Also want to look into the stuff with animations being glitchy when suddenly changing movement direction but don't really know what to look for or how to fix it.