RECOIL. General update to the ranged weapon structs with the creation of some new structs for easier organisation.
Recoil has been added for all guns currently should only be working for the player however I think I know how to get it to work for enemies by casting to the pawn class from the owner to then get the control rotation that is currently acquired using the get player controller function.

Recoil is customisable using recoil amount float, recoil duration and recoil knockback amount.

Recoil knockback is the force at which the shooter is propelled in the opposite direction of their aim after shooting.