Slide jump change. Sliding on slopes.
Slide jump no longer boosts the player forward without changing the speed of the player so they keep their velocity.

Increase to speed when sliding down slopes.
Decrease to speed when sliding up slopes. Once you stop moving whilst sliding on a slope you will go in the opposite direction back down the slope.

Started working on a slope detection for normal and sprinting movement but didn't want to do it anymore so not doing it and deleted it.