Fixed an issue with ending wall-running when touching the ground. Started working on available actions widget for HUD. Altered some things with can aim, can fire being set. Double jump change. Slide update.
Trace whilst wall running that will detect if touching the floor and will then stop you from having wall-running being active so this should fix the weird issue with the wall-running bool staying active and making you do a wall jump if you jump after wall running.

Available actions widget and available action text widget in the same style as the spedometer. Tracking of the available actions done in the event tick of the HUD component.

On landed in player now also resets their jump count for the double jumping.

Can now spin around as much as you want when sliding and it won't mess with the velocity and the checking of velocity for stopping sliding. Sliding cooldown now properly working.

More stuff in event ticks of different components for can aim and such.