Fixed some issues with interact collision overlaps for weapons. Temp audio for rocket launcher shot and rocket explosion (need to make playing the sound from the explosion more modular)
Rocket launcher. On hit explosion category of info done for ranged weapon projectiles.

Rocket launcher basically fully working with rocket exploding on impact the only thing to change is damage percentage based on proximity to the eye of the explosion. Maybe also add launching any pawn characters in the explosion radius like physics bodies are done.

Thinking of revamping the entire health system at least for the player.

Need to make the floating damage numbers a component so can use it for the explosion damage as well as normal shooting damage as well as making it easier for using with melee weapons.

Currently using temporary audio for shooting the rocket launcher and the explosion from the rockets but its ok for now.