Explosion update. Shoot trace update. Hit scan variable added for ranged weapons. Floating damage numbers implemented.
Explosion now will launch hit characters as well as the physics bodies it hits.

Shooting trace no longer default hits through objects however could have this as a piercing setting later on for ranged weapons.

New bool variable for ranged weapons for if they are hit scan basically if they should do a hit trace for hitting enemies so the rocket launcher is not a hit scan and instead its damage is focused on hitting the rocket onto an enemy.

Also updated the hit trace to use the same location as the trace for firing the projectile towards the middle of the screen / from their eyes so where they are aiming is where the hit trace will go.

Floating damage numbers now a actor component that is on the pawn that is taking damage and is now self contained so can use the floating damage in other damage sources such as the explosion from rockets which is now implemented.