Damage properly presented and done for explosion. Explosion camera shake. Piercing ranged weapons bool variable and functionality.
Floating damage numbers are now in its own component that goes onto the character.
Explosion damage is now properly represented with floating damage numbers and is also now related to the distance from the epicentre of the explosion for the damage of the ranged weapon.

Explosion from projectiles now have camera shake.

Ranged weapons can now also be piercing so that they go through objects and other characters. This is a bool variable set for ranged weapons that is then checked in the shoot component.