Custom projectile setup done for ranged weapon modification. Ranged weapon upgrade actor component created for upgrading ranged weapon info using ranged weapon info options enum. Crosshair update.
Custom projectile setup implemented for editing weapons in editor before the game started.

Upgrading a weapon uses passing ranged weapon info and the enum for the info that will be changed. Use a pure function in the upgrade ranged weapon component to change the info will need to add to this in the future if adding new info that can be changed during the game.

Now just need to make modular upgrades for the game that can be acquired but probably going to do this later on in development.

Created submachine, shotgun and rocket launcher crosshairs in aesprite and no longer using any crosshair for revolver so added functionality for hiding the crosshair if a weapon has no crosshair icon in its data table.

Also fixed an issue with the line trace for hit scan weapons hitting the other projectiles that were being shot out for multi projectile weapons like the shotgun so now keeping track of actors to ignore as an array in the shoot component which keeps track of the owner (the weapon) and the character shooting as well as the other projectiles in the shot. This array is cleared everytime the character shoots.