Melee weapon hit boxes and hit detection.
Melee weapons now have hitboxes that will detect physics bodies and apply a force to the physics body in the direction of the weapon.

New notify in melee weapon attacks which is hitbox start which then starts the hit box for the currently equipped melee weapon - will need to update this for enemies in the future since they probably won't be using the player weaponry component.

Hitboxes for melee weapons done in the event tick of the melee attack component with the tick set enabled and disabled as necessary.

When an actor is hit by the attack of a melee weapon they are added to an array of hit enemies so that the one attack doesn't hit the enemies thousands of times in one attack and then the array is cleared when the attack starts so it resets properly every time.

New float variable in the melee weapon struct for the damage of the melee weapon.

Will need to add more eventually for hit feedback for melee weapons to feel as impactful as the ranged weapons.