Melee hit feedback update.
Added some slight camera shake to hitting an enemy with a melee weapon and also a smaller camera shake to hitting a physics body with a melee weapon.

Knockback to hit character from melee weapon.

Blood splash system by making use of this tutorial: https://www.youtube.com/watch?v=vbCxZau8ag4. Adapted for use for both melee and ranged damage. Think Im going to actually try to follow the end of the tutorial to make the other blood splatter material.

Audio added and implemented for hitting an enemy with an attack for both melee and ranged and also audio for the blood actually splattering onto the environment.