The player can now have starting weapons and a starter weapon.
Had to change it from player start to having the player spawned in the world and then auto possess that but it works.

To fix an issue with the weapon display in the HUD the starting weapons equip the next weapon and then the previous weapon after setting the equipped weapon which fixes it and doesn't cause any problems as of my knowledge.