Health rework. Projectile new on hit effect - do damage. Fixed some more stuff with ammo and reloading
New take damage. New health bar and heart. One shot protection.  Setup death but not actually implemented as of yet. Healing. Health bar lerps to the target health so smooth.

New projectile on hit which will be to deal damage to the hit actor.

Floating damage stuff is now in the health take damage function instead of each individual so consoldiated it.

Fixed some more stuff with the reloading using the actual time of reloading and time to reload for more accurate sounds and stuff.