Camera update. Speed Lines.
Slight lens flare.
Some vignette for the camera.
Little bit of ambient occlusion and a little bit of motion blur.

Speed lines implemented with the help of this tutorial: https://www.youtube.com/watch?v=-HxTFzN5iMI
Niagara system attached to the player's camera and then activated and deactivated when the velocity passes a set boundary in the custom movement handler's event tick. Getting the speed lines niagara system component actor through the player interface on begin play in the custom movement handler. Its subtle but it should be.