Ground slam. Air Dash. Temp slide sound.
Implemented a ground slam ability by pressing C whilst in the air which adds an impulse downwards and then if they are still holding the ground slam button on impact with the ground a shockwave is sent out to knock up any characters or physics bodies in an area around the character. Specific ground slam niagara effect for landing which is a shockwave of the colour of the surface the character is landing on as well as some of other particles sent out similar to the footfalls system. Also a niagara for being propelled downwards by the ground slam similar to the speed lines but more exaggerated and their own system.

Air dash - dash in the air in the last movement input vector with a short cooldown that is automatically reset when the player touches the ground. Has its own sound with no niagara since it automatically activates the speed lines niagara but might make one later. Air dashing sets the movement mode to flying for a very very short time to stop gravity and to keep you in the air.

Temporarily added in sound for when sliding which is just the footfalls sound for the surface the character is sliding on but will more than likely change this in the future since it sounds kinda bad.