Unarmed attacks. Melee attack now converted from being on the weapon to being on the character.
Unarmed attacks that can be done by pressing V which can be done whilst having a weapon but are meant to be a weaker attack that is used to more so knock back and up enemies depending on the attack with different damage numbers for each unarmed attack. 3 unarmed attacks planned : uppercut, slash, and punch. Implemented the uppercut which will do some damage but send enemies high into the sky.

WHEN IMPLEMENTING NEW ATTACK ANIMATIONS MAKE SURE TO MAKE THEM LOOPING SO NO ISSUES ARISE WITH SPAMMING THE ATTACK BUTTON THIS SHOULD NOT HAVE TAKEN AS LONG BUT I DIDNT WRITE THIS DOWN OR REMEBER SO REMEMBER FOR THE FUTURE AND IT WON'T TAKE AS LONG.