Organised unarmed attack folders. Blocking.
Blocking with katana implemented.

New state machine for blocking and blocking actor component that goes onto the characters themselves.

Blocking calculation for effectiveness of damage taken in the take damage function in hte health component. Will need to juice the impact of blocking an impact from a weapon but the reduced damage taken is there. May also need to add where blocking an attack then puts the blocking cooldown on a small cooldown. Stopping blocking voluntarily will put the blocking cooldown on a small cooldown so its annoying to spawn blocking but will probably reduce this in the future. Max block time is currently 2.5 but could change dependent on the weapon so may need to add a blocking struct in and add it to the melee weapon info struct.