Perfect Block functionality.
Perfect blocking detection on take damage whilst blocking checking if the blocking effectiveness is inside the perfect block threshold. If so the character blocking takes no damage and then a shockwave. Specific sound for perfect block.

Blocking info struct which is also on the melee weapon struct so can get the blocking info from getting the melee weapon since it is part of the struct.