Melee attacking heavy attacking implemented. Melee heavy shooting projectile implemented.
Hold attack whilst having a melee weapon to start charging a melee weapon which enters a 1 frame animation state and then on release it will do a heavy attack.

Swapping weapons will stop charging the heavy attack.

Charging up the heavy attack does not increase the damage.

Custom float time for minimum charge time for the heavy attack.

Notify in heavy attack animation to shoot the projectile and using a new socket in the melee weapon mesh for where the projectile should spawn from and using the trace that is used in the shoot component to aim in the animation blueprint notify to aim the projectile.