Bloodsplosion destroy actor. Melee Beam niagara and collision.
Bloodsplosion actor will now destroy itself after a set time which is longer than the current lifetime of the niagara so if increasing the lifetime will need to increase this delay for destroying the actor.

Melee beam component improved and using a new niagara effect that also has collision and its own actor for the melee beam. Audio for the melee beam spawning and impacting in the world. May need some more work in the future.