Recoil now uses 2 curves for horizontal and vertical recoil. Starting weaponry update and possible fix.
YAYAYAYAYAYAY FINALLY.

Recoil now makes use of 2 new curves using a new function library that I set up using this unreal forum: https://forums.unrealengine.com/t/how-to-get-a-value-from-a-runtime-float-curve-structure/146059/5

We get the float value of the curves at the time for the recoil percent which is then multiplied by the recoil amount vertical and recoil amount horizontal variables respectively and then multiplied by the amount to add to the recoil percent.

This means we also created another variable in the recoil struct for the recoil amount horizontal since now recoil can be done sideways as well as vertically.

Also a new variable for the horizontal recoil direction which is set randomly from 50% chance at the start of the recoil function call.

The recoil code is now much simpler using these new variables and curves.
This took so much longer than it should have since C++ recompiling wasn't working with the original project.

Removed unused variables from recoil component as well.