Optimised animation blueprint. Recoloured all folders again.
Multithreading for the majority of the animation blueprint now.
Fast tracking the majority of the animation state blendings in the anim graph as well.
Had to create some new variables for the fast tracking.
Setting references to the components on begin play as well.
Unarmed attack info is now also set in the component so can get it threadsafe since its set at begin play for the component and probably won't be updated during runtime but if it is then just need to also update this in the component as well.