Enemy velocity damage and enemy on enemy velocity damage.
New capsule collision on the enemy parent blueprint.
New collision trace channel for the enemy velocity check which ignores weapons, projectiles, player, enemy but will overlap all things associated with non moving objects e.g. world static and world dynamic.
This capsule collision makes use of this new collision trace channel to get the velocity as the enemy hits the object and then using the hit event from the enemy capsule collision that is the main one.
Added another check in the enemy specific functionality for the take damage function in the health component to check if the damage dealer is implementing the enemy interface and if not then we update the chasing character. Also fixed an issue here since it was using the hit actor instead of the actor who did damage for updating the chasing info so this should be working properly.

2 new variables in the enemy parent blueprint which are the velocity threshold that must be reached for the velocity of the enemy for them to take damage from the movement and then the other variable is the velocity the enemy was moving at when they hit the object.

If an enemy hits another enemy when moving at very high speed the enemy they hit will also take damage according to the velocity of the enemy that hit them which will allow for interesting plays where players slingshot enemies into other enemies.