Grappling and swinging update.
Removed some previous grappling blueprints as going for a new style of having a point that is sent out by the player that they grapple towards using the mesh I created in blender. Just need to add more juice to this aspect like having it in the player's hand, getting the animations better. Having the meshes glow when in use etc. Hit stop. Camera shake.
Grappling to an moving object.
Grappling an object to the player.
Grappling the player to a stationary object.
Hold grapple button whilst having just grappled to a stationary object allows for swinging around that hit location with force being applied towards that point from current position so can increase speed fast and not boring like normal physics pendulum swinging.
Base shitty animations setup for the grappling can always change in the future.
Checks for grappling stuff in aim and attack handler so cannot aim and cannot attack whilst grappling or swinging.