Hiding Weapon whilst grappling and showing grappling mesh. Got rid of duplicated assets. Grappling mesh in hand and destroyed after stopped (still some bugs and edge cases need to find)
Currently equipped weapon is hidden whilst grappling and swinging and also any weapons picked up during grappling or swinging will be set to hidden until the grappling and swinging has stopped. This is done using functions in the weaponry component and the grappling component.

Grapple mesh in hand during grappling and swinging.

Deleted duplicated assets for the UE character.