More edge cases covered for grappling and swinging. Grappling and swinging camera shake. Turned off the check in animation blueprint for control rotation. Grappling Cable implemented. Changed more add impulses to use mass of hit. Variable delay for pulling delay for grappling. Double jump reset on grapple.
Camera shake when grappling hook attached.
Camera shake when grappling hook pulled.
Camera shake when finishing grappling.

Stopped the constant checking of control rotation in the animation blueprint since it was jittering the camera.

Cable actor setup in the player and setup in the begin play of the grappling component which is then used to attach to grappling hook. Using the same material for the grappling hook for the grappling cable and setting the length of the cable constantly whilst swinging and once when setup for the pull.

Slow time for grappling.
Slow time for swinging.