HUD shake widget animation and functionality. Swinging launch out of swinging. Blood material recreated from Reign of the Abyss.
New animation in the HUD widget that affects all other widgets inside of it and enlarges them and shears them a little bit to make the HUD react to actions in the world.

This is used in player landing, double jump, air boost, shooting, take damage, ground slam.

New launch character when finishing swinging which is affected by the velocity of the player so slower going more force added for the launch.

Blood UI material recreated from RotA and put into the game for use in the health bar and also made the heart darker in the HUD / health widget.