Slow time post process. Low Health post process. Slow time post process functionality. Fixed issue with exploding piercing projectiles.
https://www.youtube.com/watch?v=KLOGaHgo2GY
https://www.youtube.com/watch?v=_5gSoYPaDxY

New material combining work from the above 2 tutorials to create a blur radial effect and a outline with noise to exaggerate this.

Functionality in the slow time manager for the changing of the blur and the outline material as well as changing the colour of the edge noise when low on health.

Also fixed an issue with exploding projectiles that pierce so now when we overlap with an object as a piercing projectile we get the actor location as the hit location and the impact point since this was glitching out and exploding piercing rockets were exploding at 0,0,0.