Global pitch modulation functionality. Slow Time start and slow time end sounds. Slow time continuous audio component on player and functionality. Audio component specificity
Setting the global pitch modulation to be the time dilation when setting the time dilation in the slow time manager component.

Setting the global pitch modulation to be back to 1 when the time dilation returns to normal.

Play sound 2D when start slow time and the time dilation is equal to 1 meaning that slow time is starting from normal time and then this also activates the continuous audio component on the player.

Play sound 2D when the slow time ends and time dilation is set back to 1, this sound is the same as the start slow time sound but with the pitch as -1 to reverse it. The continuous audio component is deactivated here as well.