Decrease swing launch force. Tweak all can bools so don't have to have all components on everything. Removed unnecessary components from enemies. Animation blueprint can checks now based on component validity. Bloodsplosion actually heals player now. Created funky post process for tweaking how the game looks.
Swing force when ending the swing has been decreased.

Changed bools that require other less obvious components but its basically just one component that was requiring another one so this was changed. Removed all unnecessary components from enemy actor.

Changed anim blueprint checks for info to be based on whether the component associated with it is valid so not doing unnecessary checks and changing variables that don't need changing constantly even if they are multi-threaded.

Bloodsplosion trace was for enemies instead of players so thats fixed now and will properly heal players.

Made a funky little post proces just messing around with stuff trying to find a style that stands out and could be used for hte game or find multiple different variations to use for different levels.