After image material and functionality with new component.
https://www.youtube.com/watch?v=NCfAdEvHMKQ
Used the above tutorial with new functionality for the after image component.

After image activates during slow time. Has a delay that can be set. Has a after image frequency that can be set. Has a fade time that can be set as well.

Made the material be more translucent so the player can still see through them.