Optimise after image. Double Jump Slow time and Air Dash Slow time. Object Pooling for after image. After Image custom colour.
Decreased the frequency of the after images.

Double jump slow time functionality implemented.
Air dash slow time functionality implemented.

Object pooling setup for the after image actors in the after image component. Instead of constantly creating and destroying them we are instead spawning them on begin play and just setting their information when necessary and setting hidden when they would be destroyed.

The after images no longer have overlap collisions and ignore all collision types. The after image meshes no longer produce shadows as well so that's better for performance.

New vector parameter for the after image material which is the colour and is now setup in the after image component and the after image actor.