Frequently used systems now make use of the built in system pooling. Organised ground slam component. Turned off lots of ticks that weren't being used in components and components on player.
The particle systems that are very frequently like the shoot system and the projectile hit system.

Organised the ground slam component since it was a bit messy.

Turned off the tick of lots of different actors and components and then also turned off the ticks for components on the player.