Projectile pooler changed to be an actor component that will be placed on actors that can shoot. Test Area advancement. Fixed an already fixed issue that I accidently recreated in the starting weapons component. Grappling no longer spawning instead just using one. Lerping actor rotation for sliding now so smoother. Think I fixed the weird wall-running thing changed the traces a bit.
Projectile pooler is no longer an actor that is placed in the world and is instead an actor component that will be placed on all actors that will be able to shoot a weapon. This is so that we are not getting a reference to an actor in the world and instead just getting a reference to the component on the owner of the weapon. Additionally this is so that instead of having to tweak the number of projectiles spawned for every single actor in the level in this one projectile pooler actor we can instead tweak the number of projectiles spawned from each projectile pooler component dependent on the weapon that the actors will be able to shoot.

New area, deleted the previous one. 4 smaller areas within it: armoury / shooting range, indoor combat area (not started yet), swinging / grappling area, sliding / wall-running area. Had to add the inventory area back in again since deleted that without realising so also renamed the actors in the world so don't forget.

Fixed an issue again with the starting weapons component messing up the data for the starting weapons so doing some of the operations in different order.

Grappling is no longer spawning and destroying a grappling blade actor and is instead just spawning one at the start of the game and moving that one around since this should be better for garbage collection especially since the grappling blade actor doesn't really do much other than visual representation. Grappling blade also no longer casts shadows just to aid with performance a bit more.

Changing the actor rotation in the event tick for the sliding component is now changed to be lerping it so that it is a bit smoother and not as jerky.

Changed the traces for the wall-running detection quite a bit and think it may have fixed the weird opposite movement bug for the wall-running. Also doubled the wall-running speed since it was feeling really slow.