Altered health event tick so more optimised. Full-auto firing fixed with fire cooldown bool also changed burst and charge and charge burst. Continuous Charging weapons for both charging and burst charging implemented. Can fire also determined by reloading now as well. Fixed reloading and stopping aiming during shooting. Increased player gravity to 2.5.
Event tick for the health component is now only active if the player is the owner of the component. Also changed the setting of OSP active to now be a function that is checked when taking damage and when healing after the health has been updated which works as intended still and removes the constant checking from every single actor with the health component and also consolidated the check if damage broke osp function into this check osp function which is just playing a sound when osp is broke.

New bool variable called fire cooldown in attack handler. This is instead of setting can fire in the various aspects of the event graph since can fire should only be set in the event tick. Fire cooldown does dictate can fire however since its a separate variable it fixes the full-auto fire rate issue which I didn't even realise was in the game so now full-auto weapons actually fire at the displayed rate. This also fixed some of the other fire types as well since they weren't actually working properly.

New float variable in the attack handler called continuous charging time which is used to allow the continuous holding of the attack button for charging and charge burst weapons with the charge time actually still needing to be met. This variable is set after the cooldown is finished as the current triggered seconds and then when checking if the triggered seconds has surpassed the charge time we subtract it from the continuous charging time and then reset this to 0 when the button is released.

Can fire bool is now also determined by whether we are reloading as well so cannot fire whilst reloading. Also setup some stuff for charging and burst charging continuous whilst stopping aiming and whilst reloading so its all very smooth and such.

Changed some more stuff with reloading and stopping aiming whilst attempting to fire. Attempting to aim bool in aim component. Attempting to attack bool in attack handler.

Changed the player gravity to 2.5 which I think is better for the grappling and swinging and overall but it could be placebo need to change the set gravity in all instances to 2.5 but the find in blueprints thing isn't working right now.