Hip firing. Hip fire spread new float variable in ranged weapon info. Created states for all hip fire weapons in hipfire state machine. Can no longer aim whilst grappling or swinging. Unarmed attack blending with hipfire / melee ready whilst sliding and normally
Can now shoot ranged weapons from the hip without needing to be aiming.

There is now a new variable in the ranged weapon struct which is for hip firing spread.
This is setup with all of the make ranged weapon structs in various blueprints as well as in the weapon construction script for setting the weapon info.

Switch the rotation of the projectile spread and trace spread based on whether we are aiming to use the spread for hip firing or aiming.

New states for all of the hip fire state machine things. Setup with transitions but this will be changed in the animation overhaul update where I change how all / majority of the state machines work to be cohesive and easier to edit in the future.

Can no longer aim whilst grappling / swinging since aiming requires a non-offhand action to be in use.

Unarmed attack blending now with hip fire / melee ready whilst sliding and normally. This doesn't need to be done for aiming since you can't aim whilst unarmed attacking vice versa.