Lots more hipfire animations. Unarmed attacking is no longer stopped when switching weapons. Fixed issue with unarmed attack notifies looping animation. Locomotion animations fixed for layered blend per bones. Fixed issue with weapon animations in HUD from weapons that aren't originally from the player.
Hipfire animations and aim offsets completed for now for the sub, rocket, revolver (changed) and shotgun. These are put into the animation blueprint now as well

Unarmed attacking now continues whilst changing weapons.

Unarmed attacks no longer have an infinite hitbox issue with the animation looping since we are now setting the attacking false and unarmed attacking false faster than before which fixed the looping issue and fixed the animation notifies being called again.

Animation blends that use the main states for normal locomotion now uses Mesh space rotation blend which is now better for the majority of the animations. The only animation blends that don't look very good are the blocking animations and possibly the melee attacking animations.

Had to remove some of the stuff done in the ammo component for setting the HUD reference on begin play since not all weapons will do this properly since not all weapons are spawned with the player as the owner OBVIOUSLY. So this is now fixed. But if I delete the HUD component from the component and then compile the engine crashes so I'm leaving it alone for now.