New blocking animation. New melee ready animation. Changed how the interact collision radius works for weapons. Changed socket locations on guns. Changed aim at crosshair functionality in shoot component.
Katana blocking animation v2 created and implemented.
New melee ready animation for the katana created and implemented.

Interact radius for weapons now no longer set in construction script and instead just set through the interact collision without use of variable float.

Changed the sockets on all guns to be closer to the barrel so less interactions where you shoot through the floor whilst aiming downwards.

Changed the aim at crosshair functionality in the shoot component to be based on the camera's location and rotation if it is the player that is shooting so it is way more accurate for the player and then also kept in the stuff for enemies, using casting here could always change this in the future idk maybe.