Grapple fading out before being hidden. Fixed issue with grappling and spamming the button / cooldown being reset when it shouldn't. Also altered the rocket launcher shoot socket to remove hitting the player's collision when shooting at certain angle.
On begin play in the grapple blade actor we create a dynamic material instance for the inactive grapple and set it as a variable that will be used in the future.

New custom event called fade grapple which uses a timeline to lerp the opacity of the dynamic material instance to 0 from 1 and then when finished sets the grapple blade to be hidden in game. Custom event to also stop the grapple fade which just stops the timeline. Before the timeline is started we set the play rate based on a passed duration with 1/this duration. These events are called from the grapple component and helps to fade the grapple out after use rather than just disappearing.

Fixed where you could spam the grapple button whilst successfully grappling to do another grapple which would then bug out the rest of the code and the cooldown would then activate with the grapple pull only being done after this cooldown. Now the cooldown timer isn't cleared when starting or stopping swinging and is never cancelled since it doesn't need to be and so this removes this bug and is probably overall better anyway.