After Image Weapon representation. After image weapon material setup in actor and component. Fixed after image pose bug. Tested with some retro style post processes but idk. Removed some print strings. Changed base crosshair.
Static mesh child of the poseable mesh in the after image actor. This static mesh is only visible if the passed weapon static mesh is valid. Separate dynamic material instance for weapon static mesh so can set the colour separately. Pass the weapon static mesh from the after image component as well as the colour, static mesh is done by getting the basic info of the equipped weapon from the owner's weaponry component which is set in begin play. The dithering fading effect is done the same way using the same node just with the other material also plugged in now as well. The static mesh in the actor is set to attach to the poseable mesh with the socket as the Weapon_R socket.

Fixed a bug with the poseable mesh not updating the pose fast enough so you could see the previous pose for a split second so now there is a 0.01 extra delay before visibility is set to be true after setting the pose.

Changed the base crosshair that you have when no weapon equipped to be the basic crosshair.

Messed around with some retro / pixelation post processes from some youtube tutorials but probably not going to use them. Need to find an artstyle to use or some visual uniqueness for the game.