Enemy velocity damage simpler and better now using get velocity of last tick. Variable ground slam and ground landing particle scale and speed based on landing velocity
Enemy velocity damage is now using get velocity of last tick .

On land event in player changed so that we can now set the speed of the landing or ground slam particles based on how fast the player was moving as well as the scale of the particles depending on how fast the player was moving.

Also the amount of camera shake when landing is based on how fast the player was moving when they land.